Odroid C2: Fix power tree for both USB0 (OTG port) and USB1 (USB hub)
Both PHYs are internal and always supplied by the internal power rail.
The power rails of the external components are separate and have their own enable/disable/reset:
GPIOAO5 controls the power switch (U20) of the OTG port.
GPIOAO4 controls the active low reset of the USB hub (U21). The hub itself controls the enables of the dual power switch (U19) for the usb host ports.