Odroid C2: Fix power tree order to avoid -EPROBE_DEFER
The devices defined in the device tree are populated in the order of their definition in the device tree.
The three regulators regulator-tflash_vdd, gpio-regulator-tf_io, regulator-vcc1v8 are defined before their source.
This results in an -EPROBE_DEFER at boot, which is unneeded and can be avoided by moving the definitions after the definiton of their source.